+++ client/src/utils/videoHelper.js
@@ -24,24 +24,25 @@ function dataURLToBlob(dataURL)
     return new Blob([uInt8Array], { type: contentType });
 }
 
-export async function sendFrame(videoRef, canvasRef) {
+export async function sendFrame(videoRef, canvasRef, formRef) {
     if(videoRef.current.readyState === videoRef.current.HAVE_ENOUGH_DATA) {
         canvasRef.current.getContext('2d').drawImage(videoRef.current, 0, 0, 640, 480);
         const frame = canvasRef.current.toDataURL('image/jpeg', 1.0);
         const imageBlob = dataURLToBlob(frame); // Convert the frame to a Blob
 
+        const csrfToken = formRef.current.querySelector('#csrf_token').value;
+
+        console.log (csrfToken)
+
+        const formData = new FormData();
+        formData.append('image', imageBlob, 'image.jpg');
+        formData.append('csrf_token', csrfToken);
+
         try {
-            const csrfRes = await axios.get("http://127.0.0.1:3000/generate_csrf_token");
-            const csrfToken = csrfRes.data.csrf_token;
-            console.log (csrfToken)
-
-            const formData = new FormData();
-            formData.append('image', imageBlob, 'image.jpg');
-            formData.append('csrf_token', csrfToken);
-
             const response = await axios.post("http://127.0.0.1:3000/process", formData, {
                 headers: {
-                    'Content-Type': 'multipart/form-data'
+                    'Content-Type': 'multipart/form-data',
+                    'X-CSRF-TOKEN': csrfToken
                 }
             });
 
@@ -52,14 +53,19 @@ export async function sendFrame(videoRef, canvasRef) {
     }
 }
 
-export async function initVideo(videoRef, canvasRef) {
+export async function initVideo(videoRef, canvasRef, formRef) {
     if (navigator.mediaDevices && navigator.mediaDevices.getUserMedia) {
         navigator.mediaDevices.getUserMedia({ video: true })
             .then(function(stream) {
                 videoRef.current.srcObject = stream;
 
-                setInterval(() => {
-                    sendFrame(videoRef, canvasRef);
+                setInterval(async () => {
+                    const csrfRes = await axios.get("http://127.0.0.1:3000/generate_csrf_token");
+                    const csrfToken = csrfRes.data.csrf_token;
+
+                    formRef.current.querySelector('#csrf_token').value = csrfToken;
+
+                    sendFrame(videoRef, canvasRef, formRef);
                 }, 1000);             
             })
             .catch(function(err) {
